refactor app picker
