+++ apps/registry/src/components/aomi-frame.tsx
@@ -64,7 +64,7 @@ type RootProps = {
   /** Optional concrete application row to route runtime calls to. */
   applicationId?: number | string | null;
   /** Optional backend platform filter for the session app catalog. */
-  appCatalogPlatform?: string | null;
+  appPlatform?: string | null;
   /** Optional runtime client overrides. */
   clientOptions?: Omit<AomiClientOptions, "baseUrl">;
 };
@@ -109,7 +109,7 @@ const Root: FC<RootProps> = ({
   showSidebar = true,
   backendUrl,
   applicationId,
-  appCatalogPlatform,
+  appPlatform,
   clientOptions,
 }) => {
   const resolvedBackendUrl =
@@ -122,7 +122,7 @@ const Root: FC<RootProps> = ({
     <AomiRuntimeProvider
       backendUrl={resolvedBackendUrl}
       applicationId={applicationId}
-      appCatalogPlatform={appCatalogPlatform}
+      appPlatform={appPlatform}
       clientOptions={clientOptions}
     >
       <SidebarProvider className="min-h-0! h-full">
