+++ packages/client/src/index.ts
@@ -68,6 +68,12 @@ export {
 } from "./types";
 
 // =============================================================================
+// App Descriptors
+// =============================================================================
+
+export { normalizeAppDescriptor, appIdentityKey } from "./app-descriptor";
+
+// =============================================================================
 // Session (high-level orchestrated client)
 // =============================================================================
